requirements.txt
requests
readline
python-dotenv
prompt-toolkit
rich